+++ components/pm-subscription-handler/Changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 * Moved to alpine base image (DCAEGEN2-2292)
 * Added model-invariant-id and model-version-id to filter (DCAEGEN2-2151)
+* Added support for multiple CDS blueprints (DCAEGEN2-2405) 
 
 ## [1.1.0]
 ### Changed