+++ components/pm-subscription-handler/Changelog.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.2]
+### Changed
+* Bug fix for missing sdnc params in DELETE event (DCAEGEN2-2483)
+
 ## [1.1.1]
 ### Changed
 * Moved to alpine base image (DCAEGEN2-2292)